reorganize logic utilities
