+++ src/orange/CMakeLists.txt
@@ -21,6 +21,7 @@ list(APPEND SOURCES
   detail/BIHBuilder.cc
   detail/BIHPartitioner.cc
   detail/DepthCalculator.cc
+  detail/LogicUtils.cc
   detail/OrangeInputIOImpl.json.cc
   detail/RectArrayInserter.cc
   detail/SurfacesRecordBuilder.cc
@@ -49,7 +50,6 @@ list(APPEND SOURCES
   orangeinp/detail/InternalSurfaceFlagger.cc
   orangeinp/detail/IntersectSurfaceState.cc
   orangeinp/detail/LocalSurfaceInserter.cc
-  orangeinp/detail/LogicUtils.cc
   orangeinp/detail/NodeSimplifier.cc
   orangeinp/detail/ProtoBuilder.cc
   orangeinp/detail/ProtoMap.cc
