+++ src/orange/OrangeInputIO.json.cc
@@ -28,9 +28,9 @@
 
 #include "OrangeInput.hh"
 #include "OrangeTypes.hh"
-#include "orangeinp/detail/LogicUtils.hh"
 #include "surf/SurfaceTypeTraits.hh"
 
+#include "detail/LogicUtils.hh"
 #include "detail/OrangeInputIOImpl.json.hh"
 
 namespace celeritas
@@ -165,8 +165,7 @@ void from_json(nlohmann::json const& j, VolumeInput& value)
     else
     {
         // Convert logic string to vector
-        value.logic = orangeinp::detail::string_to_logic(
-            j.at("logic").get<std::string>());
+        value.logic = detail::string_to_logic(j.at("logic").get<std::string>());
         value.bbox = get_bbox(j);
     }
 }
@@ -191,7 +190,7 @@ void to_json(nlohmann::json& j, VolumeInput const& value)
     // Convert logic string to vector
     if (!value.logic.empty())
     {
-        j["logic"] = orangeinp::detail::logic_to_string(value.logic);
+        j["logic"] = detail::logic_to_string(value.logic);
     }
 
     // Write optional values
