+++ test/orange/orangeinp/CsgTreeUtils.test.cc
@@ -9,8 +9,8 @@
 #include "orange/OrangeTypes.hh"
 #include "orange/orangeinp/CsgTree.hh"
 #include "orange/orangeinp/CsgTypes.hh"
+#include "orange/orangeinp/detail/CsgLogicUtils.hh"
 #include "orange/orangeinp/detail/InternalSurfaceFlagger.hh"
-#include "orange/orangeinp/detail/LogicUtils.hh"
 #include "orange/orangeinp/detail/SenseEvaluator.hh"
 #include "orange/surf/VariantSurface.hh"
